requirements.txt
langchain
langchain_community
langchain_huggingface
faiss-cpu
pypdf
load_dotenv
streamlit